+++ CHANGELOG.md
@@ -1,6 +1,7 @@
-# 1.22.7 (2024-tba-tba)
+# 1.22.7 (2024-10-24)
 
 ## Updates/Fixes
+
 - fixed ledger trade asset bug [#427](https://github.com/asgardex/asgardex-desktop/pull/428)
 - fixed cosmos url [#413] (https://github.com/asgardex/asgardex-desktop/pull/418)
 - fixed btc/ uxto max send [#434] (https://github.com/asgardex/asgardex-desktop/pull/437)